profile_generator/requirements.txt
google-cloud-aiplatform
sqlalchemy
google-cloud-firestore
google-cloud-storage
firebase-admin
vertexai
firebase_admin